requirement.txt
streamlit
langchain
langchain-openai
langchain-core
newsapi-python
python-dotenv
